Add boot integration using the boom boot manager
Add support to optionally create snapshot boot entries and rollback
boot entries when creating snapshot sets (or later via the Manager
interface).

This introduces a new module, snapm.boot that interfaces with boom
and allows the Manager class to maintain a cache of SnapshotSet
objects to their corresponding boot entries.