snapm: fix Manager.split_snapshot_sets() with empty sources
The snapm.manager.Manger.split_snapshot_sets() method checks whether the
original snapshot set would be empty after the operation and raises an
error if this is the case. There is no corresponding check to ensure
that the destination snapshot set is not empty (i.e. the sources passed
to the split_snapshot_set() method were empty).

The snapm.command wrappers already require the sources argument:

  root@localhost:~/src/git/snapm# snapm snapset split upgrade noupgrade
  usage: snapm snapset split [-h] NAME NEW_NAME SOURCE [SOURCE ...]
  snapm snapset split: error: the following arguments are required: SOURCE

But the API does not enforce this, leading to a snapshot set with no
sources and NrSnapshots == 0:

  >>> p = m.split_snapshot_set("upgrade", "noupgrade", [])
  DEBUG - Blocking termination signals {<Signals.SIGINT: 2>, <Signals.SIGTERM: 15>}
  INFO - Attempting to split snapshot set 'upgrade'
  DEBUG - Splitting snapshot  from snapshot set 'upgrade'
  INFO - Attempting rename for snapshot set 'upgrade' to 'noupgrade'
  DEBUG - Unblocking termination signals {<Signals.SIGINT: 2>, <Signals.SIGTERM: 15>}
  >>> p
  <snapm._snapm.SnapshotSet object at 0x7eff5a90d220>
  >>> print(p)
  SnapsetName:      noupgrade
  Sources:
  NrSnapshots:      0
  Time:             2025-04-01 15:02:37
  UUID:             fab385b4-5912-54f6-8458-9b487eaf7b50
  Status:           Active
  Autoactivate:     yes
  Bootable:         no

In practice this makes the call a no-op but it's pointless and confusing
to allow this combination of arguments: raise a SnapmArgumentError if
the source_specs list is empty or None.

Fixes: a32e85f
Related: #98, #174
Resolves: #179