timers: fix build on epel-9/cs9 due to py3.10 union notation
The `Timer.__init__` method uses `x | y` union notation for the
calendarspec argument (which can be a string or a `CalendarSpec`
instance):

    class Timer:
        """
        High-level interface for managing schedling timers.
        """

        def __init__(
            self, timer_type: TimerType, name: str, calendarspec: str | CalendarSpec
        ):
            """
            Initialise a new ``Timer`` object from the provided arguments.

            :param type: A ``TimerType`` enum value representing the type of timer
                         to create.
            :param name: The name of the timer. The value is used as the instance
                         part of the timer unit name.
            :param calendarspec: A valid calendarspec string, or an instance of the
                                 ``CalendarSpec`` class representing the trigger
                                 condition for the timer.
            """

This causes the test suite to fail on epel-9/centos-stream-9 since the
shorthand union notation was introduced in py3.10:

https://docs.python.org/3/library/typing.html#special-forms

Switch to using the conventional notation using `typing.Union`:

    def __init__(
        self, timer_type: TimerType, name: str, calendarspec: Union[str, CalendarSpec]
    ):

This is compatible with all supported Python versions. The problem was
not picked up at the time fcae54da / #254 was merged since nothing at
the time actually imported the timers module. This became apparent with
commit 0246fd5 which imports the `Schedule` interface (which in turn
imports `Timer`).

Fixes: fcae54da
Resolves: #311